# Conflicts:
#	src/Raven.Client/project.json
#	src/Raven.Server/Documents/DocumentsStorage.cs
#	src/Raven.Server/Documents/Versioning/VersioningStorage.cs
#	test/FastTests/Client/Attachments/AttachmentsVersioning.cs
#	test/FastTests/Server/Documents/Versioning/VersioningTest.cs
#	test/Tests.Infrastructure/RavenTestBase.cs
#	test/Tryouts/Program.cs

+++ test/Tests.Infrastructure/RavenTestBase.cs
@@ -15,6 +15,7 @@ using Raven.Client.Documents.Session;
 using Raven.Client.Extensions;
 using Raven.Client.Server;
 using Raven.Client.Server.Operations;
+using Raven.Server;
 using Raven.Server.Config;
 using Raven.Server.Config.Attributes;
 using Raven.Server.Documents;
@@ -46,10 +47,16 @@ namespace FastTests
             Action<DatabaseDocument> modifyDatabaseDocument = null,
             Func<string, string> modifyName = null,
             string apiKey = null,
-            bool ignoreDisabledDatabase = false)
+            bool ignoreDisabledDatabase = false,
+            int replicationFacotr = 1,
+            RavenServer defaultServer = null,
+            bool waitForDatabasesToBeCreated= false,
+            bool deleteDatabaseWhenDisposed = true,
+            bool createDatabase = true)
         {
             lock (_getDocumentStoreSync)
             {
+                defaultServer = defaultServer??Server;
                 var name = caller != null
                     ? $"{caller}_{Interlocked.Increment(ref _counter)}"
                     : Guid.NewGuid().ToString("N");
@@ -83,38 +90,45 @@ namespace FastTests
                     int.MaxValue.ToString();
                 modifyDatabaseDocument?.Invoke(doc);
 
-                TransactionOperationContext context;
-                using (Server.ServerStore.ContextPool.AllocateOperationContext(out context))
+                if (createDatabase)
                 {
-                    context.OpenReadTransaction();
-                    if (Server.ServerStore.Read(context, Constants.Documents.Prefix + name) != null)
-                        throw new InvalidOperationException($"Database '{name}' already exists");
+                    TransactionOperationContext context;
+                    using (defaultServer.ServerStore.ContextPool.AllocateOperationContext(out context))
+                    {
+                        context.OpenReadTransaction();
+                        if (defaultServer.ServerStore.Cluster.Read(context, Constants.Documents.Prefix + name) != null)
+                            throw new InvalidOperationException($"Database '{name}' already exists");
+                    }
                 }
 
                 var store = new DocumentStore
                 {
-                    Url = UseFiddler(Server.WebUrls[0]),
+                    Url = UseFiddler(defaultServer.WebUrls[0]),
                     DefaultDatabase = name,
                     ApiKey = apiKey
                 };
                 ModifyStore(store);
                 store.Initialize();
 
-            store.Admin.Server.Send(new CreateDatabaseOperation(doc));
+                if (createDatabase)
+                    store.Admin.Server.Send(new CreateDatabaseOperation(doc, replicationFacotr));
+                
+
                 store.AfterDispose += (sender, args) =>
                 {
                     if (CreatedStores.TryRemove(store) == false)
                         return; // can happen if we are wrapping the store inside sharded one
 
-                    if (Server.Disposed == false)
+                    if (defaultServer.Disposed == false)
                     {
-                        var databaseTask = Server.ServerStore.DatabasesLandlord.TryGetOrCreateResourceStore(name, ignoreDisabledDatabase);
+                        var databaseTask = defaultServer.ServerStore.DatabasesLandlord.TryGetOrCreateResourceStore(name, ignoreDisabledDatabase);
                         if (databaseTask != null && databaseTask.IsCompleted == false)
                             databaseTask.Wait();
-                                // if we are disposing store before database had chance to load then we need to wait
+                        // if we are disposing store before database had chance to load then we need to wait
 
-                        Server.Configuration.Server.AnonymousUserAccessMode = AnonymousUserAccessModeValues.Admin;
-                    store.Admin.Server.Send(new DeleteDatabaseOperation(name, hardDelete));
+                        defaultServer.Configuration.Server.AnonymousUserAccessMode = AnonymousUserAccessModeValues.Admin;
+                        if(deleteDatabaseWhenDisposed)
+                            store.Admin.Server.Send(new DeleteDatabaseOperation(name, hardDelete));
                     }
                 };
                 CreatedStores.Add(store);
